+++ demo/ajax.php
@@ -2,6 +2,6 @@
 
 include 'bootstrap.php';
 $debugbar['messages']->addMessage('hello from ajax');
-$debugbar->sendDataInHeaders();
+$debugbar->sendDataInHeaders(true);
 ?>
 hello from AJAX